requirements.txt
Django
django-cors-headers
djangorestframework
channels
channels-redis
Pillow
python-dotenv
gunicorn